[cont-android] feat: diálogo inteligente para cantidades (fraccionables) en compras y ventas (#43)
* [cont-android] feat: diálogo inteligente para cantidades en compras y ventas

* feat: ajustar incremento Android y añadir inventario web + respaldo cifrado API

* [cont-free] feat: separar inventario en módulo independiente con sidebar

* [cont-free] feat: sidebar de 2 módulos y pestañas inferiores por hoja

+++ sysgd-cont/android/app/src/main/java/cu/lazaroysr96/sysgdcont/data/model/InventarioModels.kt
@@ -48,7 +48,7 @@ data class LineaVenta(
     val productoId: String,
     val nombreProducto: String,
     val precioUnitario: Double,
-    val cantidad: Int
+    val cantidad: Double
 ) {
     val subtotal: Double get() = precioUnitario * cantidad
 }
@@ -103,7 +103,7 @@ data class LineaCompra(
     val productoId: String,
     val nombreProducto: String,
     val precioUnitario: Double,
-    val cantidad: Int
+    val cantidad: Double
 ) {
     val subtotal: Double get() = precioUnitario * cantidad
 }
